requirements.txt
langchain
streamlit
streamlit_chat
chromadb
pypdf
fastembed
langchain_community
Pillow